use new result format for frequentist
